Move getRepos function to contex
