+++ src/App.js
@@ -17,20 +17,6 @@ const App = () => {
 	const [loading, setLoading] = useState(false);
 	const [alert, setAlert] = useState(null);
 
-	// Get user repositories
-	const getUserRepos = async login => {
-		setLoading(true);
-
-		const res = await axios.get(
-			`https://api.github.com/users/${login}/repos?per_page=5&sort=created:asc&client_id=${
-				process.env.REACT_APP_GITHUB_CLIENT_ID
-			}&client_secret=${process.env.REACT_APP_GITHUB_CLIENT_SECRET}`
-		);
-
-		setRepos(res.data);
-		setLoading(false);
-	};
-
 	// Lift alert state
 	const showAlert = (msg, type) => {
 		setAlert({ msg, type });
@@ -58,19 +44,7 @@ const App = () => {
 								)}
 							/>
 							<Route exact path='/about' component={About} />
-							<Route
-								exact
-								path='/user/:login'
-								render={props => (
-									<Fragment>
-										<User
-											{...props}
-											getUserRepos={getUserRepos}
-											repos={repos}
-										/>
-									</Fragment>
-								)}
-							/>
+							<Route exact path='/user/:login' component={User} />
 						</Switch>
 					</div>
 				</div>
