+++ src/context/github/GithubState.js
@@ -55,7 +55,21 @@ const GithubState = props => {
 		});
 	};
 
-	// Get repos
+	// Get user repositories
+	const getUserRepos = async login => {
+		setLoading();
+
+		const res = await axios.get(
+			`https://api.github.com/users/${login}/repos?per_page=5&sort=created:asc&client_id=${
+				process.env.REACT_APP_GITHUB_CLIENT_ID
+			}&client_secret=${process.env.REACT_APP_GITHUB_CLIENT_SECRET}`
+		);
+
+		dispatch({
+			type: GET_REPOS,
+			payload: res.data
+		});
+	};
 
 	// Clear users
 	const clearUsers = () => dispatch({ type: CLEAR_USERS });
@@ -71,7 +85,8 @@ const GithubState = props => {
 				loading: state.loading,
 				searchUsers,
 				clearUsers,
-				getUser
+				getUser,
+				getUserRepos
 			}}
 		>
 			{props.children}
