+++ src/context/github/githubReducer.js
@@ -31,6 +31,12 @@ export default (state, action) => {
 			user: action.payload,
 			loading: false
 		};
+	case GET_REPOS:
+		return {
+			...state,
+			repos: action.payload,
+			loading: false
+		};
 	default:
 		return state;
 	}
